requirements/base.txt
certifi==2018.4.16
chardet==3.0.4
coreapi
coreschema
Django==2.2
django-cors-headers==3.5.0
django-rest-swagger==2.1.2
django-summernote==0.8.8.6
django-templated-mail==1.1.1
djangorestframework
djangorestframework-jsonapi
djangorestframework_simplejwt
djoser
idna==2.7
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
psycopg2==2.8.5
PyJWT==1.6.4
pytz==2018.5
requests==2.19.1
simplejson==3.16.0
uritemplate==3.0.0
urllib3==1.23
python-dotenv==0.10.3